Eliminado de usuario
Arreglos de malas prácticas e implementación de eliminado de usuario.
