+++ users.h
@@ -4,14 +4,15 @@
 #include <string>
 #include <fstream>
 #include <vector>
+using namespace std;
 
 class User{
-public: 
+public:
 	std::string name; //nombre del usuario
 	float score; //puntuación del usuario
 public:
 	User();
-	void setname(); //setter del nombre de usuario
+	void setname(string ); //setter del nombre de usuario
 	std::string getname(); //getter del nombre de usuario
 	void setscore(float); //setter del score del usuario
 	float getscore(); //getter de la puntuación del usuario
